Add *probable* const values to fields
I don't think there's a method to return the const value, but for
enums it *seems* to be the case that the const value for enums is equal
to the order they are returned when enumerating fields.  The 'Offset'
column on the class form fields tab will now show that index
as the probably const value, i.e. 'Const: 0'